feat(srvc-inventory):added one service per Repository with simple register method

+++ repo-inventory/src/main/java/fr/lateb/domain/entity/EntryEntity.java
@@ -5,5 +5,5 @@ import fr.lateb.type.BeerStatus;
 import java.util.Date;
 import java.util.UUID;
 
-public record EntryEntity(UUID id, Boolean forSale, Float buyPrice, Date DDM, Date saleDate, Integer order,
+public record EntryEntity(UUID id, Boolean forSale, Float buyPrice, Date DDM, Date saleDate, Long order,
                           Long beer, BeerStatus status) {}
