+++ repo-inventory/src/main/java/fr/lateb/domain/entity/OrderEntity.java
@@ -2,4 +2,4 @@ package fr.lateb.domain.entity;
 
 import java.util.Date;
 
-public record OrderEntity(Integer id, String supplier, Date arrivalDate) {}
+public record OrderEntity(Long id, String supplier, Date arrivalDate) {}
